+++ src/raytracer/hittable/box.h
@@ -17,11 +17,12 @@ namespace raytracer
 	{
 	public:
 		Box() = default;
+
 		Box(const point3& p0, const point3& p1, const ref<Material>& ptr);
 
-		bool hit(const Ray& r, double t_min, double t_max, hit_record& rec) const override;
+		bool hit(const Ray& r, float t_min, float t_max, hit_record& rec) const override;
 
-		bool boundingBox(double time0, double time1, Aabb& output_box) const override;
+		bool boundingBox(float time0, float time1, Aabb& output_box) const override;
 
 	public:
 		point3 box_min{};
