+++ src/raytracer/hittable/hit_record.h
@@ -7,16 +7,17 @@
 namespace raytracer
 {
 	class Ray;
+
 	class Material;
 
 	struct hit_record
 	{
 		point3 p;
 		vec3 normal;
-		double t;
+		float t;
 		bool front_face;
 		ref<Material> mat_ptr;
-		double u, v;
+		float u, v;
 
 		void setFaceNormal(const Ray& r, const vec3& outward_normal);
