+++ src/raytracer/hittable/hittablelist.cpp
@@ -9,14 +9,16 @@
 namespace raytracer
 {
 
-	bool HittableList::hit(const Ray& r, double t_min, double t_max, hit_record& rec) const
+	bool HittableList::hit(const Ray& r, float t_min, float t_max, hit_record& rec) const
 	{
 		hit_record temp_rec{};
 		bool hit_anything = false;
 		auto closest_so_far = t_max;
 
-		for (const auto& object : m_objects) {
-			if (object->hit(r, t_min, closest_so_far, temp_rec)) {
+		for (const auto& object: m_objects)
+		{
+			if (object->hit(r, t_min, closest_so_far, temp_rec))
+			{
 				hit_anything = true;
 				closest_so_far = temp_rec.t;
 				rec = temp_rec;
@@ -41,14 +43,15 @@ namespace raytracer
 		add(object);
 	}
 
-	bool HittableList::boundingBox(double time0, double time1, Aabb& output_box) const
+	bool HittableList::boundingBox(float time0, float time1, Aabb& output_box) const
 	{
 		if (m_objects.empty()) return false;
 
 		Aabb temp_box;
 		bool first_box = true;
 
-		for (const auto& object : m_objects) {
+		for (const auto& object: m_objects)
+		{
 			if (!object->boundingBox(time0, time1, temp_box)) return false;
 			output_box = first_box ? temp_box : Aabb::surroundingBox(output_box, temp_box);
 			first_box = false;
