+++ src/raytracer/textures/noisetexture.cpp
@@ -6,13 +6,13 @@
 
 namespace raytracer
 {
-	color NoiseTexture::value(double u, double v, const point3& p) const
+	color NoiseTexture::value(float u, float v, const point3& p) const
 	{
-		return color(1,1,1) * 0.5 * (1 + sin(m_scale*p.z + 10*m_noise.turb(p)));
+		return color(1, 1, 1) * 0.5f * (1 + sin(m_scale * p.z + 10 * m_noise.turb(p)));
 	}
 
-	NoiseTexture::NoiseTexture(double sc)
-	: m_scale(sc)
+	NoiseTexture::NoiseTexture(float sc)
+			: m_scale(sc)
 	{
 
 	}
